Function query abstraction; Simplied get routes
